Updated project. Added confirmation checking. Updated readme. Added license file.

+++ src/app/lib/view_models/home/home_view_model.dart
@@ -13,6 +13,7 @@ class HomeViewModel extends BaseViewModel {
     this._tagsManager,
     this._notesManager,
     this._navigationService,
+    this._dialogService,
   ) {
     currentAccount = accountManager.currentAccount;
     tags = [];
@@ -27,6 +28,7 @@ class HomeViewModel extends BaseViewModel {
   final TagsManager _tagsManager;
   final NotesManager _notesManager;
   final NavigationService _navigationService;
+  final DialogService _dialogService;
 
   StreamSubscription _tagsStreamSubscription;
   StreamSubscription _notesStreamSubscription;
@@ -97,9 +99,9 @@ class HomeViewModel extends BaseViewModel {
 
   Future<void> onShowSettings() => _navigationService.push(ViewNames.settingsView);
 
-  Future<void> onAddNote() => _navigationService.pushModal(ViewNames.addNoteView);
+  Future<void> onAddNote() => _navigationService.pushModal(ViewNames.noteView);
 
-  void onEditTags() => _navigationService.push(ViewNames.editTagsView);
+  void onEditTags() => _navigationService.push(ViewNames.tagsView);
 
   void onToggleEditingMode() {
     switch (editingMode) {
@@ -140,21 +142,31 @@ class HomeViewModel extends BaseViewModel {
     }
   }
 
-  Future<void> onViewNote(NoteItemModel note) => _navigationService.pushModal(ViewNames.addNoteView, parameter: note);
+  Future<void> onViewNote(NoteItemModel note) => _navigationService.pushModal(ViewNames.noteView, parameter: note);
 
   void Function() get onDeleteSelectedNotes {
     if (selectedNotesCount == 0)
       return null;
     else {
       return () async {
-        final notesToBeDeleted = notes.where((n) => n.isSelected);
-        for (final note in notesToBeDeleted) {
-          await _notesManager.deleteNote(NoteEntity(id: note.id));
-        }
+        final shouldDelete = await _dialogService.confirm('This cannot be undone.', title: 'Delete note/s?', ok: 'Delete');
 
-        onToggleEditingMode();
+        if (shouldDelete == true) {
+          final notesToBeDeleted = notes.where((n) => n.isSelected);
+          for (final note in notesToBeDeleted) {
+            await _notesManager.deleteNote(NoteEntity(id: note.id));
+          }
+
+          onToggleEditingMode();
+        }
       };
     }
+  }
+
+  Future<bool> onBackPressed() async {
+    final isDeleting = editingMode == ListEditingMode.delete;
+    if (isDeleting) onToggleEditingMode();
+    return !isDeleting;
   }
 
   void _onFilterTags(TagItemModel tag) {
@@ -173,16 +185,24 @@ class HomeViewModel extends BaseViewModel {
         id: n.id,
         title: n.title,
         content: n.content,
+        created: n.created,
         lastModified: n.lastModified,
         tag: n.tag,
       );
-    }).toList();
+    }).toList()
+      ..sort((n1, n2) {
+        if (n2.lastModified == null) {
+          return n2.created.compareTo(n1.lastModified ?? n1.created);
+        } else {
+          return n2.lastModified.compareTo(n1.lastModified ?? n1.created);
+        }
+      });
 
     selectedNotesCount = notes.where((n) => n.isSelected).length;
   }
 
   void _onTagsChanged(List<TagEntity> newTags) {
-    final newTagModels = newTags.map((t) => TagItemModel(t.id, t.name, t.created)).toList();
+    final newTagModels = newTags.map((t) => TagItemModel(t.id, t.name, t.created, t.lastModified)).toList();
     final selectedTagExists = newTagModels.contains((t) => t.id == selectedTag?.id);
 
     if (selectedTagExists) {
